+++ source/Private/Get-StartMenuFolderEventLogSecurityDescriptor.ps1
@@ -1,0 +1,7 @@
+function Get-StartMenuFolderEventLogSecurityDescriptor {
+    [CmdletBinding()]
+    [OutputType([string])]
+    param()
+
+    'O:BAG:SYD:(A;;0x7;;;SY)(A;;0x7;;;BA)(A;;0x3;;;IU)'
+}